requirements.txt
aiogram==3.16.0
aiohttp>=3.10,<3.12
aiosqlite==0.20.0
python-dotenv==1.0.1
yookassa==2.5.0
